% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{fv_strict}
\alias{fv_strict}
\title{Helper function with parameter check}
\usage{
fv_strict(a, s = 2, e = 1e-04)
}
\arguments{
\item{a}{numeric}

\item{s}{numeric}

\item{e}{numeric thr}
}
\description{
Helper function with parameter check
}
